feat(CfnCallback) Provide possibility to callback cloud formation
This change enables performing a callback to cloud formation. By design this function does not act as a custom resource. Instead it is a helper for building a custom resource that can orchestrate Redshift actions.

In order to make Cloudformation failures easier to troubleshoot the finished event that is returned will be annotated with an error if execution failed.